LangGraphAgentRAG/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bitarray==3.3.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
dnspython==2.7.0
docx2txt==0.8
environs==14.1.0
faiss-cpu==1.9.0.post1
fastapi==0.115.8
fastapi-offline==1.7.3
fastapi_cors==0.0.6
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
gritql==0.1.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-cli==0.0.35
langchain-community==0.3.17
langchain-core==0.3.35
langchain-ollama==0.2.3
langchain-postgres==0.0.13
langchain-text-splitters==0.3.6
langgraph==0.3.5
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
langserve==0.3.1
langsmith==0.3.8
lxml==5.3.1
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.4.7
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pgvector==0.3.6
propcache==0.2.1
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg-pool==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pybloom_live==4.0.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
sse-starlette==1.8.2
starlette==0.45.3
tabulate==0.9.0
tenacity==9.0.0
tomlkit==0.13.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
